Exercise 2

+++ assignment_improvement_exercise (2).py
@@ -4,7 +4,7 @@
 
 import math
 
-radius = int(input("please enter the radius of the circle: "))
+radius = float(input("please enter the radius of the circle: "))
 
 circumference = 2* math.pi * radius
 circumference = round(circumference,2)
